Move EmailTemplate model to proper location
