+++ src/core/models/v1/EmailTemplate.py
